CommandAttribute now doesn't accept names with spaces in them.
Some more unit tests. This time finishing up YACCS.Commands.Linq.Parameters.

Fixed an issue with GetParametersByType doing the type checking backwards.

DefaultValue on IImmutableParameter/IParameter no longer relies on DBNull, instead using a simple object. Added HasDefaultValue to their interfaces to accommodate that; it has a setter, but that setter only does something if the value is false, and in that case it sets DefaultValue to the NoDefaultValue object.