Unit tests for TypeReaderRegistry.
Invalid readers for types now throw when adding and not when retrieving.

Attemping to add in a nullable type reader for a value type now needs the generic interface implemented.

ITypeReader<T>#ReadAsync now returns an ITask<T> instead of a Task<T> so it can covariantly return.

Added in the package MorseCode.ITask to handle the above.