Rewrote Args.TryParse to be less "bruteforcey" in parsing.
Instead of making 2 initial passes through the input and storing all start/end quote positions in 2 separate lists there is only 1 initial pass through the input to gather extrema for start/end and the expected size of the output array.

Then when actually parsing instead of using a list an array can be used since we now get the expected size on the intial pass.