Moved ParameterPreconditions into the Preconditions namespace.
Added in the basis for Groups of preconditions, there is currently not an actual implementation for handling the groups in the CommandService yet.

CommandService now returns an ICommandResult when failing which provides the command (if there is one) and the parameter (if there is one). For example, now if a parameter precondition fails the parameter name can be displayed.

Successfully executing a command has not been changed because that will always have a non null Command and will never have a parameter.