Tidied up some of CommandScore creation and changed the static methods to extension methods in a new namespace.
Added in the ability to lower the priority of a generated named argument command.

Added in more tests to ensure generated named arguments commands and normal commands return errors which make sense with what a user probably was invoking.

In YACCS.Examples, added in Console.ReleaseIOLocks to the delay and sleep command because otherwise they only get released when the context is disposed after the delay/sleep.